+++ crates/application/src/favorites/favorite_transfer.rs
@@ -22,7 +22,7 @@ use vrcx_0_application_core::WebClient;
 use vrcx_0_vrchat_client::http_api::parse_api_json;
 
 const FAVORITE_RECOVERED_GROUP: &str = "Recovered";
-const FAVORITE_TRANSFER_PAGE_SIZE: i64 = 300;
+const FAVORITE_TRANSFER_PAGE_SIZE: i32 = 300;
 const FAVORITE_TRANSFER_MAX_PAGES: usize = 50;
 const FAVORITE_TRANSFER_REMOTE_INTERVAL: Duration = Duration::from_millis(250);
 
@@ -814,7 +814,7 @@ async fn fetch_online_favorite_index(
     let equivalent_types = kind_equivalent_favorite_types(kind);
     let mut by_object_id = HashMap::new();
     let mut group_counts: HashMap<String, i64> = HashMap::new();
-    let mut offset = 0_i64;
+    let mut offset = 0_i32;
 
     for _ in 0..FAVORITE_TRANSFER_MAX_PAGES {
         let request = favorites_get_input(
