+++ crates/application/src/social/group_calendar.rs
@@ -23,7 +23,7 @@ use vrcx_0_vrchat_client::http_api::{ApiJsonResponse, ApiScope, HttpApiRequestIn
 
 use crate::{Error, Result};
 
-const CALENDAR_PAGE_SIZE: i64 = 100;
+const CALENDAR_PAGE_SIZE: i32 = 100;
 const CALENDAR_MAX_PAGES: usize = 50;
 const GROUP_PROFILE_CONCURRENCY: usize = 4;
 const GROUP_PROFILE_CACHE_CAPACITY: u64 = 32;
@@ -182,7 +182,7 @@ async fn collect_calendar_pages(
     let mut rows = Vec::new();
     for page_index in 0..=CALENDAR_MAX_PAGES {
         ensure_scope_matches(&deps.auth_scope, scope)?;
-        let offset = (page_index as i64) * CALENDAR_PAGE_SIZE;
+        let offset = (page_index as i32) * CALENDAR_PAGE_SIZE;
         let params = CalendarListParams {
             n: Some(CALENDAR_PAGE_SIZE),
             offset: Some(offset),
